# InnoSync-Mobile-App/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ApplicationResponse:
    properties:
      applied_at:
        type: string
      cover_letter:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      project_role_id:
        type: integer
      project_title:
        type: string
      role_name:
        type: string
      status:
        $ref: '#/definitions/models.ApplicationStatusEnum'
      updated_at:
        type: string
      user_full_name:
        type: string
      user_id:
        type: integer
    type: object
  models.ApplicationStatusEnum:
    enum:
    - PENDING
    - UNDER_REVIEW
    - ACCEPTED
    - REJECTED
    - WITHDRAWN
    type: string
    x-enum-varnames:
    - ApplicationStatusPending
    - ApplicationStatusUnderReview
    - ApplicationStatusAccepted
    - ApplicationStatusRejected
    - ApplicationStatusWithdrawn
...
  models.CreateApplicationRequest:
    properties:
      cover_letter:
        type: string
    type: object
...
  models.UpdateApplicationStatusRequest:
    properties:
      status:
        $ref: '#/definitions/models.ApplicationStatusEnum'
    required:
    - status
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      last_login:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for InnoSync Mobile Application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: InnoSync API
  version: "1.0"
paths:
  /applications:
    get:
      description: Get all applications submitted by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ApplicationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user applications
      tags:
      - applications
  /applications/{id}:
    delete:
      description: Withdraw a submitted application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete application
      tags:
      - applications
    get:
      description: Get a specific application by ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApplicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get application
      tags:
      - applications
  /applications/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a specific application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateApplicationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApplicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update application status
      tags:
      - applications
  /applications/project-roles/{projectRoleId}:
    get:
      description: Get all applications for a specific project role
      parameters:
      - description: Project Role ID
        in: path
        name: projectRoleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ApplicationResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get role applications
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Apply for a project role
      parameters:
      - description: Project Role ID
        in: path
        name: projectRoleId
        required: true
        type: integer
      - description: Application creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ApplicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create application
      tags:
      - applications
...
  /auth/me:
    get:
      description: Get information about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"